code review and secure deletion of a post (can delete only the person whose post)

+++ posts-service/src/main/java/org/example/postsservice/service/PostsService.java
@@ -69,15 +69,17 @@ public class PostsService {
 
     @Transactional
     public void deletePost(DeletePostDto postDto) {
+        Optional<Profile> author = profileRepository.findByUsername(postDto.getUsername());
+        if (author.isEmpty()) return;
         Optional<Post> postToDeleteOptional = postRepository.findById(postDto.getPostId());
         if (postToDeleteOptional.isPresent()) {
             Post postToDelete = postToDeleteOptional.get();
+            if (!postToDelete.getAuthor().getId().equals(author.get().getId())) return;
             postToDelete.setAuthor(null);
             postRepository.save(postToDelete);
             shareSubscribersClient.deleteFromSubscribers(postToDelete.getId());
         }
     }
-
 
     private static Post unwrapPost(PostDto post) {
         Post postToUnwrap = new Post();
